See http://code.google.com/p/solr-php-client/issues/detail?id=1
Changed require_once statements that expected the Solr Client library to be on the include path into statements that use the __FILE__ magic constant to do them as absolute paths. This should get rid of the requirement for the user to change their include path configuration or add the library to the include path at runtime.

Provided as a convenience for the user.